refactor(svar2): extract the shared view selection predicate into svar2_view
